Do not use renamed symbols when performing symbol name surgery.
This is necessary because otherwise it will take the gensym-style
numeric suffix of a renamed identifier and use that as the basis for
the new symbol.  In cases where you perform low-level surgery on the
symbol string representation, you almost always want to strip syntax.

This actually got broken because of CHICKEN core changeset
41d48cdba2e46645c7dd157525652aa3953aed6c, which fixed bug #518.  This
has an effect on some small set of macros that have always been
incorrect, but which would just happen to work in some situations.